Simplify organisation + more docs
