--- rs/canister/examples/greet/canister/src/api/insert_jokes.rs
@@ -1,14 +1,0 @@
-use candid::CandidType;
-use serde::{Deserialize, Serialize};
-use std::collections::HashMap;
-
-#[derive(CandidType, Serialize, Deserialize)]
-pub struct Args {
-    pub jokes: HashMap<u32, String>,
-}
-
-#[derive(CandidType, Serialize, Deserialize)]
-pub enum Response {
-    Success(u32),
-    NotAuthorized,
-}
