--- rs/canister/examples/greet/canister/src/updates/mod.rs
@@ -1,1 +1,0 @@
-pub mod insert_jokes;
