--- rs/canister/examples/reminder/src/lifecycle/mod.rs
@@ -1,5 +1,0 @@
-mod init;
-mod post_upgrade;
-mod pre_upgrade;
-
-const READER_WRITER_BUFFER_SIZE: usize = 1024 * 1024; // 1MB
