+++ rs/offchain/examples/discord/src/openchat.rs
@@ -77,7 +77,7 @@ pub async fn start_openchat_bot(
     let routes = Router::new()
         .route("/execute_command", post(execute_command))
         .route_layer(ExtractJwtLayer::new())
-        .route("/", get(bot_definition))
+        .fallback(get(bot_definition))
         .layer(
             ServiceBuilder::new()
                 .layer(TraceLayer::new_for_http())
