+++ rs/sdk/src/api/mod.rs
@@ -1,5 +1,2 @@
-mod api_key_registry;
 pub mod command;
 pub mod definition;
-
-pub use api_key_registry::ApiKeyRegistry;
