+++ rs/sdk/src/api_key_registry.rs
