+++ rs/sdk/src/lib.rs
@@ -1,9 +1,9 @@
 pub mod api;
+mod api_key_registry;
 pub mod mainnet;
 pub mod oc_api;
-mod runtime;
 pub mod types;
 mod utils;
 
-pub use runtime::*;
+pub use api_key_registry::ApiKeyRegistry;
 pub use utils::*;
