+++ rs/sdk/src/oc_api/client_factory.rs
@@ -1,4 +1,4 @@
-use crate::runtime::Runtime;
+use super::Runtime;
 use crate::types::ActionContext;
 use client::Client;
 use std::sync::Arc;
