+++ rs/sdk/src/oc_api/client_factory/client/create_channel.rs
@@ -1,8 +1,8 @@
 use super::Client;
 use crate::oc_api::actions::create_channel::*;
 use crate::oc_api::actions::ActionArgsBuilder;
+use crate::oc_api::Runtime;
 use crate::types::{AccessGateConfig, CanisterId, ChatPermissions, Document, Milliseconds, Rules};
-use crate::Runtime;
 use std::sync::Arc;
 
 pub struct CreateChannelBuilder<R> {
