+++ rs/sdk/src/oc_api/mod.rs
@@ -1,3 +1,7 @@
+mod runtime;
+
 pub mod actions;
 pub mod client_factory;
 pub mod get_access_token;
+
+pub use runtime::Runtime;
